requirements/google-colab-requirements.txt
dask==2021.9.0
distributed==2021.9.0
numba>=0.53.1
jsonschema>=3.2.0
simplejson>=3.17.2
cryptography>=3.2.1
imgkit>=1.2.2
packaging>=20.9
requests>=2.24.0
tqdm>=4.51.0
fastnumbers>=3.1.0
multipledispatch>=0.6.0
numpy>=1.21.2
yellowbrick
deprecated>=1.2.12
setuptools>=50.3.1
Jinja2>=3.0.1
pygments>=2.2.0
ipython>=7.24.1
humanize>=3.9.0
psutil>=5.8.0
ordered-set>=4.0.2
deepdiff>=5.5.0
statsmodels>=0.12.2
glom>=20.11.0
fastavro>=1.4.1
fast-histogram>=0.9
python-magic>=0.4.15; sys_platform != 'win32'
python-magic-bin==0.4.14; sys_platform == 'win32'
nltk>=3.6.2
pendulum>=2.1.2
rply>=0.7.8
pybigquery>=0.10.1
pandavro>=1.6.0
openpyxl>=3.0.7
tabulate>=0.8.9
matplotlib>=3.4.2
seaborn>=0.11.1
wordninja==2.0.0
jellyfish>=0.8.7
Metaphone==0.6
num2words==0.5.10
xlrd>=2.0.1
s3fs>=2021.7.0
aiobotocore[boto3]>=1.3.3
lxml==4.6.3
hi-dateinfer
hi-urlparser